refactored approach
