+++ pages/_document.js
@@ -16,33 +16,6 @@ class MyDocument extends Document {
             rel="stylesheet"
           />
         </Head>
-        <script
-          type="text/javascript"
-          dangerouslySetInnerHTML={{
-            __html: `      
-            var chatbox = document.getElementById('fb-customer-chat');
-            chatbox.setAttribute("page_id", "1665091083765739");
-            chatbox.setAttribute("attribution", "biz_inbox");`,
-          }}
-        />
-        <script
-          type="text/javascript"
-          dangerouslySetInnerHTML={{
-            __html: `window.fbAsyncInit = function() {
-              FB.init({
-                xfbml: true,
-                version: 'v17.0'
-              });
-            };
-            (function(d, s, id) {
-              var js, fjs = d.getElementsByTagName(s)[0];
-              if (d.getElementById(id)) return;
-              js = d.createElement(s); js.id = id;
-              js.src = 'https://connect.facebook.net/en_US/sdk/xfbml.customerchat.js';
-              fjs.parentNode.insertBefore(js, fjs);
-            }(document, 'script', 'facebook-jssdk'));`,
-          }}
-        />
         <body>
           <Main />
           <NextScript />
