+++ pages/services/[service-type]/[service-id]/index.tsx
@@ -10,6 +10,7 @@ import {
 import Layout from "src/template/page-wrapper";
 import { Service } from "src/types/service";
 import Button from "src/components/button";
+import { FacebookMessengerChat } from "src/components/fb-messenger-chat";
 import { braidsAndTwistsServices, dreadlocksServices, strings } from "data";
 import { bookingMachine } from "src/booking/booking-machine";
 import { useBookingService } from "src/helpers";
@@ -62,6 +63,7 @@ const ServiceID = ({ service, serviceTypeName }: Props) => {
       {showModal && (
         <AddingExtensions setShowModal={setShowModal} service={service} />
       )}
+      <FacebookMessengerChat />
     </Layout>
   );
 };
